backend/requirements.txt
Django~=5.2
djangorestframework~=3.15
drf-spectacular~=0.27
django-environ~=0.11
httpx~=0.27
psycopg2-binary~=2.9
pika~=1.3

# Testing
pytest~=8.3
pytest-django~=4.8
